imports.py
import nltk
from nltk.tokenize import casual_tokenize
nltk.download('averaged_perceptron_tagger')

import pandas as pd
import numpy as np

import psycopg2
import spacy
from spacy.matcher import PhraseMatcher
nlp = spacy.load("en_core_web_lg")

import re

import warnings
warnings.filterwarnings('ignore')

import seaborn as sns
import matplotlib.pyplot as plt

print("package imports done")
