requirements.txt
blis==0.4.1
catalogue==0.0.8
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
en-core-web-lg==2.2.5
idna==2.8
importlib-metadata==0.23
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==7.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.3
seaborn==0.9.0
six==1.13.0
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
tqdm==4.39.0
urllib3==1.25.7
wasabi==0.4.0
zipp==0.6.0
